优化grpc性能

+++ joyrpc-test/joyrpc-test-transport/src/test/java/io/joyrpc/transport/netty4/mock/MockMessage.java
@@ -53,7 +53,7 @@ public class MockMessage implements Message<MockMessageHeader, byte[]> {
     }
 
     @Override
-    public int getMsgId() {
+    public long getMsgId() {
         return getHeader().getMsgId();
     }
